refactor: rename test package to match source
