+++ src/test/java/org/trebol/jpa/services/predicates/CustomersPredicateJpaServiceTest.java
@@ -1,9 +1,8 @@
-package org.trebol.jpa.services.predicate;
+package org.trebol.jpa.services.predicates;
 
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Predicate;
 import org.junit.jupiter.api.Test;
-import org.trebol.jpa.services.predicates.SalespeoplePredicateJpaServiceImpl;
 
 import java.util.HashSet;
 import java.util.List;
@@ -12,12 +11,12 @@ import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class SalespeoplePredicateJpaServiceTest {
+class CustomersPredicateJpaServiceTest {
 
   @Test
   void parses_map() {
     Predicate emptyPredicate = new BooleanBuilder();
-    SalespeoplePredicateJpaServiceImpl service = instantiate();
+    CustomersPredicateJpaServiceImpl service = instantiate();
     List<Predicate> predicates = List.of(emptyPredicate,
                                          service.parseMap(Map.of("id", "1")),
                                          service.parseMap(Map.of("idNumber", "id test")),
@@ -34,7 +33,7 @@ class SalespeoplePredicateJpaServiceTest {
     assertEquals(predicates.size(), distinctPredicates.size());
   }
 
-  private SalespeoplePredicateJpaServiceImpl instantiate() {
-    return new SalespeoplePredicateJpaServiceImpl();
+  private CustomersPredicateJpaServiceImpl instantiate() {
+    return new CustomersPredicateJpaServiceImpl();
   }
 }
