+++ src/test/java/org/trebol/jpa/services/predicates/ImagesPredicateJpaServiceTest.java
@@ -1,9 +1,8 @@
-package org.trebol.jpa.services.predicate;
+package org.trebol.jpa.services.predicates;
 
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Predicate;
 import org.junit.jupiter.api.Test;
-import org.trebol.jpa.services.predicates.SellStatusesPredicateJpaServiceImpl;
 
 import java.util.HashSet;
 import java.util.List;
@@ -12,21 +11,23 @@ import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class SellStatusesPredicateJpaServiceTest {
+class ImagesPredicateJpaServiceTest {
 
   @Test
   void parses_map() {
     Predicate emptyPredicate = new BooleanBuilder();
-    SellStatusesPredicateJpaServiceImpl service = instantiate();
+    ImagesPredicateJpaServiceImpl service = instantiate();
     List<Predicate> predicates = List.of(emptyPredicate,
                                          service.parseMap(Map.of("id", "1")),
-                                         service.parseMap(Map.of("name", "name test")),
-                                         service.parseMap(Map.of("nameLike", "name portion")));
+                                         service.parseMap(Map.of("code", "code test")),
+                                         service.parseMap(Map.of("filename", "filename test")),
+                                         service.parseMap(Map.of("codeLike", "code portion")),
+                                         service.parseMap(Map.of("filenameLike", "filename portion")));
     Set<Predicate> distinctPredicates = new HashSet<>(predicates);
     assertEquals(predicates.size(), distinctPredicates.size());
   }
 
-  private SellStatusesPredicateJpaServiceImpl instantiate() {
-    return new SellStatusesPredicateJpaServiceImpl();
+  private ImagesPredicateJpaServiceImpl instantiate() {
+    return new ImagesPredicateJpaServiceImpl();
   }
 }
