+++ src/test/java/org/trebol/jpa/services/predicates/SalespeoplePredicateJpaServiceTest.java
@@ -1,9 +1,8 @@
-package org.trebol.jpa.services.predicate;
+package org.trebol.jpa.services.predicates;
 
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Predicate;
 import org.junit.jupiter.api.Test;
-import org.trebol.jpa.services.predicates.CustomersPredicateJpaServiceImpl;
 
 import java.util.HashSet;
 import java.util.List;
@@ -12,12 +11,12 @@ import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class CustomersPredicateJpaServiceTest {
+class SalespeoplePredicateJpaServiceTest {
 
   @Test
   void parses_map() {
     Predicate emptyPredicate = new BooleanBuilder();
-    CustomersPredicateJpaServiceImpl service = instantiate();
+    SalespeoplePredicateJpaServiceImpl service = instantiate();
     List<Predicate> predicates = List.of(emptyPredicate,
                                          service.parseMap(Map.of("id", "1")),
                                          service.parseMap(Map.of("idNumber", "id test")),
@@ -34,7 +33,7 @@ class CustomersPredicateJpaServiceTest {
     assertEquals(predicates.size(), distinctPredicates.size());
   }
 
-  private CustomersPredicateJpaServiceImpl instantiate() {
-    return new CustomersPredicateJpaServiceImpl();
+  private SalespeoplePredicateJpaServiceImpl instantiate() {
+    return new SalespeoplePredicateJpaServiceImpl();
   }
 }
