+++ src/test/java/org/trebol/jpa/services/predicates/ShippersPredicateJpaServiceTest.java
@@ -1,9 +1,8 @@
-package org.trebol.jpa.services.predicate;
+package org.trebol.jpa.services.predicates;
 
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Predicate;
 import org.junit.jupiter.api.Test;
-import org.trebol.jpa.services.predicates.BillingCompaniesPredicateJpaServiceImpl;
 
 import java.util.HashSet;
 import java.util.List;
@@ -12,23 +11,21 @@ import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class BillingCompaniesPredicateJpaServiceTest {
+class ShippersPredicateJpaServiceTest {
 
   @Test
   void parses_map() {
     Predicate emptyPredicate = new BooleanBuilder();
-    BillingCompaniesPredicateJpaServiceImpl service = instantiate();
+    ShippersPredicateJpaServiceImpl service = instantiate();
     List<Predicate> predicates = List.of(emptyPredicate,
                                          service.parseMap(Map.of("id", "1")),
-                                         service.parseMap(Map.of("idNumber", "id test")),
                                          service.parseMap(Map.of("name", "name test")),
-                                         service.parseMap(Map.of("idNumberLike", "id number portion")),
                                          service.parseMap(Map.of("nameLike", "name portion")));
     Set<Predicate> distinctPredicates = new HashSet<>(predicates);
     assertEquals(predicates.size(), distinctPredicates.size());
   }
 
-  private BillingCompaniesPredicateJpaServiceImpl instantiate() {
-    return new BillingCompaniesPredicateJpaServiceImpl();
+  private ShippersPredicateJpaServiceImpl instantiate() {
+    return new ShippersPredicateJpaServiceImpl();
   }
 }
